Add support for 'vaulted' shells
If a user is currently in a [vaulted](github.com/miquella/vaulted)
shell, they will now have an indication of this in the agnoster prompt,
not unlike the virtualenv prompt. The foreground color will switch if
the timer expires (which usually signifies that the temporary Amazon
keys generated by vaulted will be invalid).